+++ database/migrations/2018_11_21_062127_create_orders_products_table.php
@@ -17,10 +17,10 @@ class CreateOrdersProductsTable extends Migration
             $table->increments('id');
             $table->integer('quantify');
 
-            $table->smallInteger("order_id")->unsigned();
+            $table->integer("order_id")->unsigned();
             $table->foreign("order_id")->references('id')->on("orders");
 
-            $table->smallInteger("product_id")->unsigned();
+            $table->integer("product_id")->unsigned();
             $table->foreign("product_id")->references('id')->on("products");
 
             $table->timestamps();
